app/requirements.txt
streamlit
streamlit-drawable-canvas
torch
torchvision
pillow
psycopg2-binary
pandas